Remove redirect / reauthorize approach and replace with kill access.
Previously, the reauthorization approach would logout the live spammer, but
this didn't deter the user from simply re-registering a new account and
spamming again.

This commit changes the strategy to kill off access to the rest of the site
if a logged-in user is marked as a spammer.

This makes it harder for the live spammer to re-register to the site.